Move domain class to v3/domain package
